requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
arxiv-checker==1.4.4
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.1
bert-score==0.3.9
bleu==0.3
blis==0.4.1
bs4==0.0.1
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
cymem==2.0.3
efficiency==0.4
filelock==3.0.12
flatbuffers==1.12
fsspec==2021.7.0
future==0.18.2
gast==0.4.0
google-auth==1.19.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
huggingface-hub==0.0.12
idna==2.10
inflect==5.3.0
joblib==0.16.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.0
mecab-python3==0.996.5
multidict==5.1.0
murmurhash==1.0.2
moverscore
nltk==3.5
#numpy==1.21.0
oauthlib==3.1.0
opt-einsum==3.3.0
#packaging==20.4
pandas==1.1.1
Pillow==8.3.1
plac==1.1.3
portalocker==1.7.1
preshed==3.0.2
protobuf==3.12.2
pyaml==21.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
pyemd==0.5.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-lightning==1.4.2
pytorch-metric-learning==0.9.99
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
sacrebleu==1.4.12
sacremoses==0.0.43
scikit-learn==0.24.2
scipy==1.5.2
seaborn==0.11.2
sentencepiece==0.1.91
six==1.15.0
soupsieve==2.0.1
spacy==2.3.2
srsly==1.0.2
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
tf-slim==1.1.0
thinc==7.4.1
threadpoolctl==2.2.0
tokenizers==0.10.3
#torch==1.9.0
torchmetrics==0.5.0
#torchvision==0.6.1+cu92
tqdm==4.48.0
transformers==4.9.1
typing==3.7.4.3
typing-extensions==3.7.4.3
urllib3==1.25.10
wasabi==0.8.0
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.3
